Pass through Windows env vars in local run
For all systems `bk local run` already passes through the `HOME` environment variable so that things such as configurations (in `$HOME/.config`), caches (`$HOME/.cache`), etc can be appropriately resolved by processes that run under the local `buildkite-agent`.

On Windows however `HOME` is not used (unless you are in WSL or other Unix-like emulator like `git-bash`). Instead there are a few environment variables that should be used instead:
- `HOMEPATH` / `USERPROFILE` which point to the real "home" directory where documents, etc are stored.
- `APPDATA` where machine-global configurations and app data is stored.
- `LOCALAPPDATA` where user-specific configurations and app data is stored.

We should pass these along as well to the `buildkite-agent` on start.