Convert GitHub auth process from "web application flow" to "device flow"
When this app was originally written, the web application flow was the
only way to collect a GitHub oauth token for the user.

However, in July 2020 GitHub added support for the device flow [1]. The
device flow is nicer because only the oauth client ID needs to be
distributed in the binary - there's no need for an oauth secret any
more.

GitHub docs for the device flow are here at [2].

[1] https://github.blog/changelog/2020-07-27-oauth-2-0-device-authorization-flow/
[2] https://docs.github.com/en/free-pro-team@latest/developers/apps/authorizing-oauth-apps#device-flow