Make BUILDKITE_PIPELINE_DEFAULT_BRANCH configurable
Because `BUILDKITE_PIPELINE_DEFAULT_BRANCH` is *not configurable* through an
environment variable, I think a special purpose env var for `buildkite-cli` may
make sense. As such, I defined the `BUILDKITE_CLI_` prefix for this case. I'm
open to alternatives.

Also defaulting to `main` instead of `master`, since this is the new GH default.