Update the authoritative URLs for emoji data
In August 2021 we renamed the development branch in the emojis repo from
master to main.

That was relatively smooth from an emojis development perspective, but
we didn't realise the cli had master hard coded.

As a quick 'fix' I've re-pushed a master branch to the emojis repo,
based on current main. It won't stay up to date though, so this changes
the cli to check the main branch instead.

See also: https://github.com/buildkite/emojis/issues/386