upgrade go-git from v4 to v5
In #206 gopkg.in/src-d/go-git.v4 was added as a dependency, which has
led to a couple of GH security alerts.

Reading the details on the two alerts, it seems like we're unlikely to
be impacted. We're using it to parse git URLs and not clone repositories
onto a filesystem, so we're not working with untrusted input.

I thought I'd try upgrading the dependency to v5 anyway though. We'll be
on the latest version which will help with keeping the project
maintainable, and maybe the security alerts will go away as a side
effect.

I had to bump the default 1 min timeout on golint to get the linting
step to pass. Looking at some recent builds golint was taking very close
to 1min, so I guess this was enough to push it over? It only takes a few
seconds on my laptop, maybe we need better/bigger CI instances?